fix(ci): satisfy pr3 shape checks
